chore: rename modules with the same name as their parent
