+++ src/sourcegen/syntax.rs
@@ -5,13 +5,13 @@
 //! Generates the [`SyntaxKind`] enum and the mapping of the [`TokenKind`] enum
 
 #[cfg(test)]
-pub mod syntax {
+pub mod lib {
     use heck::ToUpperCamelCase;
     use proc_macro2::{Ident, TokenStream};
     use quote::{format_ident, quote};
 
-    use crate::sourcegen::data::data::SYNTAX_NODES;
-    use crate::sourcegen::data::data::TOKENS;
+    use crate::sourcegen::data::lib::SYNTAX_NODES;
+    use crate::sourcegen::data::lib::TOKENS;
     use crate::sourcegen::lib::{
         add_preamble, guarantee_file_content, project_path, rustfmt_content,
     };
